feat(tokens): export type, moving token file
