+++ packages/components/cascader-panel/index.ts
@@ -10,5 +10,6 @@ const _CascaderPanel = CascaderPanel as SFCWithInstall<typeof CascaderPanel>
 
 export default _CascaderPanel
 export const ElCascaderPanel = _CascaderPanel
+
 export * from './src/types'
 export * from './src/config'
