+++ packages/components/checkbox/index.ts
@@ -12,3 +12,5 @@ export default ElCheckbox
 
 export const ElCheckboxButton = withNoopInstall(CheckboxButton)
 export const ElCheckboxGroup = withNoopInstall(CheckboxGroup)
+
+export * from './src/checkbox.type'
