+++ packages/components/collapse/index.ts
@@ -8,3 +8,5 @@ export const ElCollapse = withInstall(Collapse, {
 })
 export default ElCollapse
 export const ElCollapseItem = withNoopInstall(CollapseItem)
+
+export * from './src/collapse.type'
