+++ packages/components/date-picker/index.ts
@@ -11,3 +11,5 @@ _DatePicker.install = (app: App) => {
 
 export default _DatePicker
 export const ElDatePicker = _DatePicker
+
+export * from './src/date-picker.type'
