+++ packages/components/radio/src/useRadio.ts
@@ -1,11 +1,13 @@
 import { ref, computed, inject } from 'vue'
-import { elFormKey, elFormItemKey } from '@element-plus/tokens'
+import { elFormKey, elFormItemKey, radioGroupKey } from '@element-plus/tokens'
 import { useGlobalConfig } from '@element-plus/utils/util'
-import radioGroupKey from './token'
 
 import type { ComputedRef, WritableComputedRef } from 'vue'
-import type { ElFormContext, ElFormItemContext } from '@element-plus/tokens'
-import type { RadioGroupContext } from './token'
+import type {
+  ElFormContext,
+  ElFormItemContext,
+  RadioGroupContext,
+} from '@element-plus/tokens'
 
 export const useRadio = () => {
   const ELEMENT = useGlobalConfig()
