+++ packages/components/tabs/src/tabs.type.ts
