+++ packages/components/tree-v2/index.ts
@@ -3,3 +3,5 @@ import TreeV2 from './src/tree.vue'
 
 export const ElTreeV2 = withInstall(TreeV2)
 export default ElTreeV2
+
+export * from './src/types'
