+++ packages/components/upload/index.ts
@@ -11,3 +11,5 @@ const _Upload = Upload as SFCWithInstall<typeof Upload>
 
 export default _Upload
 export const ElUpload = _Upload
+
+export * from './src/upload.type'
