+++ packages/tokens/select-v2.ts
@@ -1,6 +1,9 @@
-import type { OptionProps, SelectProps } from './defaults'
 import type { ExtractPropTypes, InjectionKey } from 'vue'
-import type { Option } from './select.types'
+import type { Option } from '@element-plus/components/select-v2/src/select.types'
+import type {
+  OptionProps,
+  SelectProps,
+} from '@element-plus/components/select-v2/src/defaults'
 
 export interface SelectV2Context {
   props: ExtractPropTypes<typeof SelectProps>
@@ -10,7 +13,8 @@ export interface SelectV2Context {
   onKeyboardSelect: () => void
 }
 
-export const selectV2InjectionKey =
-  'ElSelectV2Injection' as any as InjectionKey<SelectV2Context>
+export const selectV2InjectionKey: InjectionKey<SelectV2Context> = Symbol(
+  'ElSelectV2Injection'
+)
 export type IOptionProps = ExtractPropTypes<typeof OptionProps>
 export type ISelectProps = ExtractPropTypes<typeof SelectProps>
